refactor: wrap WebCLI in a component with hooks to get model data
